feat(papermag): remove utils/papermagSubscribe and add fireGqlRequest and getPaymentDataOfPapermagSubscription into api/newwbpay-papermag
